pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sf-music-calendar"
version = "0.1.6"
description = "Multi-venue music event scraper for San Francisco"
authors = [
    {name = "Griffin Tarpenning", email = "gtarpenning@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.0.0",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
music = "cli:main"

[tool.setuptools]
py-modules = ["cli", "main", "music_calendar", "venues_config"]

[tool.setuptools.packages.find]
include = ["models*", "scrapers*", "storage*", "ui*", "config*", "tests*"]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "build",
    "twine",
] 